Adds new VOD card Refactors Schedule card to pass icon to Card Details Refactors CardDetails to accept an icon
