+++ src/components/Card/index.js
@@ -1,4 +1,5 @@
 import Content from './ContentCard';
 import Schedule from './ScheduleCard';
+import VodCard from './VodCard';
 
-export default { Content, Schedule };
+export default { Content, Schedule, VodCard };
